# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: DriveRAG API
  version: "0.2.0"
  description: |
    Agentic Knowledge Graph RAG system API.
    This is the contract that the backend implements and the frontend consumes.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.dingdong-rag.com/api/v1
    description: Production
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: Register a new user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: login
      summary: Authenticate user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get current authenticated user
      tags: [Auth]
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
    patch:
      operationId: updateCurrentUser
      summary: Update current user profile
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
        "409":
          description: Email already in use

  # ── Documents ─────────────────────────────────────────────────────────
  /documents:
    post:
      operationId: uploadDocument
      summary: Upload a document
      tags: [Documents]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Document uploaded and queued for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResponse"
        "413":
          description: File too large
        "415":
          description: Unsupported file type
    get:
      operationId: listDocuments
      summary: List documents
      tags: [Documents]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
...
  /documents/{document_id}:
    get:
      operationId: getDocument
      summary: Get document details
      tags: [Documents]
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "404":
          description: Document not found
    delete:
      operationId: deleteDocument
      summary: Delete a document
      tags: [Documents]
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Document deleted
        "404":
          description: Document not found
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        full_name:
          type: string
          minLength: 1
          maxLength: 255
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
    UserResponse:
      type: object
      required: [id, email, full_name, is_active, is_admin, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        is_active:
          type: boolean
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [user, tokens]
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
        tokens:
          $ref: "#/components/schemas/TokenResponse"

    # ── Documents ─────────────────────────────────────────────────────
    DocumentUploadResponse:
      type: object
      required: [id, filename, file_type, file_size, status, uploaded_at]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_type:
          type: string
        file_size:
          type: integer
        status:
          type: string
          enum: [uploaded, processing, ready, failed]
        uploaded_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
    DocumentResponse:
      type: object
      required: [id, filename, file_type, file_size, status, source, chunk_count, is_base_knowledge, uploaded_at]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_type:
          type: string
        file_size:
          type: integer
        status:
          type: string
          enum: [uploaded, processing, ready, failed]
        source:
          type: string
        chunk_count:
          type: integer
        error_message:
          type: string
          nullable: true
        is_base_knowledge:
          type: boolean
        metadata:
          type: object
        uploaded_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    DocumentListResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentResponse"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...